* develop: (128 commits)
  Bump version to 1.0.47
  Hotfix: "pyyaml" version
  CTX-4915: Reversed centroid calculation
  Bump version to 1.0.46
  CTX-4587: Add "pyyaml" package to Pipfile and pyproject.toml
  CTX-4587: experiment.config -> task.yaml
  CTX-4915: FIxed bracket placement
  CTX-4915: Made centroid calculation flexible to n points
  Bump version to 1.0.45
  Hotfix: Fix bug when fetching entities
  Bump version to 1.0.44
  CTX-4915: Fixed centroid celculation
  CTX-4793: Added copyright header to functions.py
  CTX-4793: Moved functions related code to top-level module "functions" and added documentation
  CTX-4333: Updated documentation for submitOutput and submitOutputs
  CTX-4793: Moved request functions success and badRequest to networking.utils
  Bump version to 1.0.43
  CTX_4333: Updated sending request to use new code
  CTX-4333: Changed endpoint to /model-queue/output-parameter
  Bump urllib3 from 2.0.6 to 2.0.7
  ...

# Conflicts:
#	coretex/entities/dataset/network_dataset.py
#	coretex/networking/network_object.py

+++ coretex/bioinformatics/ctx_qiime2/utils.py
@@ -17,34 +17,38 @@
 
 from typing import List
 from pathlib import Path
+from zipfile import ZipFile
 
 import logging
+import shutil
 import gzip
 
-from ...coretex import Experiment, CustomSample, CustomDataset
+from ...entities import TaskRun, CustomSample, CustomDataset
+from ... import folder_manager
 
 
-def createSample(name: str, datasetId: int, path: Path, experiment: Experiment, stepName: str, retryCount: int = 0) -> CustomSample:
+def createSample(name: str, datasetId: int, path: Path, taskRun: TaskRun, stepName: str, retryCount: int = 0) -> CustomSample:
     sample = CustomSample.createCustomSample(name, datasetId, str(path))
     if sample is None:
         if retryCount < 3:
             logging.info(f">> [Coretex] Retry count: {retryCount}")
-            return createSample(name, datasetId, path, experiment, stepName, retryCount + 1)
+            return createSample(name, datasetId, path, taskRun, stepName, retryCount + 1)
 
         raise ValueError(">> [Coretex] Failed to create sample")
 
-    experiment.createQiimeArtifact(f"{stepName}/{name}", path)
+    taskRun.createQiimeArtifact(f"{stepName}/{name}", path)
 
     return sample
 
 
-def compressGzip(source: Path, destination: Path) -> None:
+def compressGzip(source: Path, destination: Path, deleteSource: bool = False) -> None:
     logging.info(f"{source} -> {destination}")
 
     with gzip.open(destination, "w") as destinationFile:
         destinationFile.write(source.read_bytes())
 
-    source.unlink()
+    if deleteSource:
+        source.unlink()
 
 
 def sampleNumber(sample: CustomSample) -> int:
@@ -54,19 +58,20 @@ def sampleNumber(sample: CustomSample) -> int:
 def isFastqMPSample(sample: CustomSample) -> bool:
     sample.unzip()
 
-    for path in sample.load().folderContent:
-        if path.name != "sequences":
-            continue
-
-        return (path / "sequences.fastq").exists() and (path / "barcodes.fastq").exists()
-
-    return False
+    return sample.joinPath("forward.fastq").exists() and sample.joinPath("barcodes.fastq").exists()
 
 
 def isFastqDPSample(sample: CustomSample) -> bool:
     sample.unzip()
 
     return any([path.suffix == ".fastq" for path in sample.load().folderContent])
+
+
+def isImportedSample(sample: CustomSample) -> bool:
+    sample.unzip()
+
+    sampleContent = [path.name for path in sample.load().folderContent]
+    return "multiplexed-sequences.qza" in sampleContent
 
 
 def isDemultiplexedSample(sample: CustomSample) -> bool:
@@ -99,12 +104,6 @@ def isPhylogeneticTreeSample(sample: CustomSample) -> bool:
     )
 
 
-def isMetadataSample(sample: CustomSample) -> bool:
-    sample.unzip()
-
-    return any([path.suffix == ".tsv" for path in sample.load().folderContent])
-
-
 def getFastqMPSamples(dataset: CustomDataset) -> List[CustomSample]:
     # Multiplexed fastq data
     return dataset.getSamples(isFastqMPSample)
@@ -113,6 +112,10 @@ def getFastqMPSamples(dataset: CustomDataset) -> List[CustomSample]:
 def getFastqDPSamples(dataset: CustomDataset) -> List[CustomSample]:
     # Demultiplexed fastq data
     return dataset.getSamples(isFastqDPSample)
+
+
+def getImportedSamples(dataset: CustomDataset) -> List[CustomSample]:
+    return dataset.getSamples(isImportedSample)
 
 
 def getDemuxSamples(dataset: CustomDataset) -> List[CustomSample]:
@@ -127,9 +130,31 @@ def getPhylogeneticTreeSamples(dataset: CustomDataset) -> List[CustomSample]:
     return dataset.getSamples(isPhylogeneticTreeSample)
 
 
-def getMetadataSample(dataset: CustomDataset) -> CustomSample:
-    metadataSample = dataset.getSamples(isMetadataSample)
-    if len(metadataSample) != 1:
-        raise ValueError(f">> [Coretex] Dataset must contain exaclty one metadata sample. Found {len(metadataSample)}")
+def getMetadata(sample: CustomSample) -> Path:
+    metadataPathList = list(sample.path.glob("*.tsv"))
+    if len(metadataPathList) != 1:
+        raise RuntimeError(f">> [Coretex] Metadata sample must contain one .tsv file. Found {len(metadataPathList)}")
 
-    return metadataSample[0]
+    return metadataPathList[0]
+
+
+def isPairedEnd(sample: CustomSample) -> bool:
+    # In order to determine whether we are dealing with paired-end
+    # sequences, this function unzips the qiime artifact and
+    # reads the metadata, looking for the second (type) row, which will have
+    # "PairedEnd" somewhere if it's paired-end
+
+    sampleTemp = folder_manager.createTempFolder("qzaSample")
+    qzaPath = list(sample.path.iterdir())[0]
+
+    with ZipFile(qzaPath, "r") as qzaFile:
+        qzaFile.extractall(sampleTemp)
+
+    metadataPath = list(sampleTemp.rglob("*metadata.yaml"))[0]
+
+    with metadataPath.open("r") as metadata:
+        pairedEnd = "PairedEnd" in metadata.readlines()[1]
+
+    shutil.rmtree(sampleTemp)
+
+    return pairedEnd
